refactor: move rerun constants into dimos/visualization/rerun/config.py
Delete dimos/visualization/constants.py and consolidate all rerun
defaults/types into rerun/config.py per review feedback (no diluted
ownership). Also wire web_port through to rr.serve_web_viewer() so it
actually sets the port, and guard Blueprint import behind TYPE_CHECKING.
